RAG_code/RAG_Simple/requirements.txt
langchain
langchain_community
streamlit
gradio
chromadb
termcolor
python-dotenv
pymupdf
sentence-transformers
fastapi
uvicorn
python-multipart
pydantic
ollama 